coderizebanking-master/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-cors-headers==3.11.0
django-sslserver==0.22
django-tenants==3.4.2
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
psycopg2==2.9.10
pycryptodome==3.14.1
pytz==2024.2
requests==2.27.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
